Revert to int8: float32 was slower on Zen+ (4.2s vs 3.8s)

+++ listener.py
@@ -56,7 +56,7 @@ oww = WakeModel(wakeword_models=[OWW_MODEL])
 whisper = WhisperModel(
     "Systran/faster-distil-whisper-small.en",
     device="cpu",
-    compute_type="float32",
+    compute_type="int8",
     cpu_threads=4,
     num_workers=2,
 )
